+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Changes on `main` will be listed here.
 
+### Bug Fixes
+
+-   [[#2](https://github.com/diatche/react-native-form-model/pull/2)] Fixed a bug when a a NaN number was set as the value.
+
 ## 0.2.0
 
 11 Jun 2021